requirements.txt
# Multi-Agent AI Dashboard - Dependencies

# Agent Orchestration
crewai>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.20

# Ollama Integration
langchain-ollama>=0.1.0
ollama>=0.1.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
chardet>=5.2.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
kaleido>=0.2.1
statsmodels>=0.14.0

# PDF Generation (ReportLab works on Windows without GTK)
reportlab>=4.0.0
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
aiofiles>=23.2.0
httpx>=0.26.0